Move the dog probe contract to shep-core
VERSION_FLAG, SHEP_PROTOCOL_KEY and the shep-protocol answer parser
were private to shep-cli's adopt, hand-copied by every dog author from
a docs snippet. A typo in that snippet reads as protocol unknown with
nothing saying so, and --schema would repeat the bug on day one.

shep_core::dogs now owns the flag names, the answer grammar and
DogVersion (plain data with no CLI concern in it, so it moved with the
parser rather than splitting the shape from the function that builds
it), plus SCHEMA_FLAG and SECRET_KEY, both unused until Task 3 and
Task 2 give them a caller. shep-cli imports from there instead of
defining its own copies; the asker itself (spawning the binary,
applying the timeout, judging the answer) stays in shep-cli beside the
rest of adopt's vetting, per decision 5.

Behaviour-neutral: no shep-cli test needed editing, only the import.
Two small unit tests moved to shep-core's dogs.rs alongside the parser
they exercise.

cargo test --workspace --lib --bins --all-features -- --skip ::slow::
2232 passed, 0 failed.
